Algo/requirements.txt
# Core Django packages
Django>=4.2.0,<5.0
django-extensions>=3.2.0
django-cors-headers>=4.3.0

# Database
psycopg2-binary>=2.9.0
django-environ>=0.11.0

# API Framework
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0

# Real-time capabilities
channels>=4.0.0
channels-redis>=4.1.0
daphne>=4.0.0

# Background tasks
celery>=5.3.0
django-celery-beat>=2.5.0
django-celery-results>=2.5.0
flower>=2.0.0

# Caching and Redis
redis>=5.0.0
django-redis>=5.4.0
hiredis>=2.2.0

# Production server
gunicorn>=21.2.0
whitenoise>=6.5.0

# Monitoring and metrics
django-debug-toolbar>=4.0.0
django-silk>=5.0.0
prometheus-client>=0.17.0
sentry-sdk[django]>=1.32.0

# Analytics and charting
numpy>=1.24.0
pandas>=2.0.0
plotly>=5.15.0

# Testing
pytest>=7.4.0
pytest-django>=4.5.0
pytest-cov>=4.1.0
factory-boy>=3.3.0

# Development tools
black>=23.7.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.5.0
django-stubs>=4.2.0

# Production utilities
django-storages>=1.14.0
boto3>=1.28.0  # For AWS S3 if needed
django-health-check>=3.17.0

# Security
django-csp>=3.7.0
django-ratelimit>=4.1.0
